change test to check scrolling
